Feat/user data

+++ src/main/java/com/spring/ApiSystem/personal/PersonalRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonalRepository extends JpaRepository <Personal, Integer>{
+public interface PersonalRepository extends JpaRepository <Personal, Long>{
 }
